+++ src/main/java/de/codescape/jira/plugins/scrumpoker/workflow/StartSessionWorkflowFunctionFactory.java
@@ -10,8 +10,7 @@ import java.util.Map;
  * Factory class for the {@link StartSessionWorkflowFunction}. This factory class is mainly empty because the function
  * does not require any configuration options yet.
  */
-public class StartSessionWorkflowFunctionFactory extends AbstractWorkflowPluginFactory
-    implements WorkflowPluginFunctionFactory {
+public class StartSessionWorkflowFunctionFactory extends AbstractWorkflowPluginFactory implements WorkflowPluginFunctionFactory {
 
     @Override
     public Map<String, ?> getDescriptorParams(Map<String, Object> map) {
@@ -20,14 +19,17 @@ public class StartSessionWorkflowFunctionFactory extends AbstractWorkflowPluginF
 
     @Override
     protected void getVelocityParamsForInput(Map<String, Object> map) {
+        // intentionally left blank
     }
 
     @Override
     protected void getVelocityParamsForEdit(Map<String, Object> map, AbstractDescriptor abstractDescriptor) {
+        // intentionally left blank
     }
 
     @Override
     protected void getVelocityParamsForView(Map<String, Object> map, AbstractDescriptor abstractDescriptor) {
+        // intentionally left blank
     }
 
 }